Finishing the DeleteMotorcycleCommandService

+++ RentalService/src/Application/Services/Commands/DeleteMotorcycleCommandService.cs
@@ -2,16 +2,21 @@
 using Application.Services.Commands.Interfaces;
 using Common.Application;
 using Domain.Repositories;
-using Microsoft.AspNetCore.Http.HttpResults;
 
 namespace Application.Services.Commands
 {
     public class DeleteMotorcycleCommandService : IDeleteMotorcycleCommandService
     {
         private readonly IMotorcycleRepository _motorcycleRepository;
+        private readonly IRentalRepository _rentalRepository;
 
-        public DeleteMotorcycleCommandService(IMotorcycleRepository motorcycleRepository)
+        public DeleteMotorcycleCommandService
+            (
+            IMotorcycleRepository motorcycleRepository,
+            IRentalRepository rentalRepository
+            )
         {
+            _rentalRepository = rentalRepository;
             _motorcycleRepository = motorcycleRepository;
         }
 
@@ -19,12 +24,13 @@ namespace Application.Services.Commands
         {
             //Verificar se existe a moto
             var motorcycle = await _motorcycleRepository.GetByIdAsync(dto.Id);
-
             if (motorcycle == null)
                 throw new BusinessException($"Moto com o id: {dto.Id} não encontrada");
-
-            //TODO: Implementar a validação de Historico de Aluguel da moto:
+         
             //Se tiver registro de locação Não remover!
+            var rental = await _rentalRepository.GetRentalByMotorcycleId(motorcycle.Id);
+            if(rental != null)
+                throw new BusinessException($"Erro ao excluir moto, ela possui historico de aluguel");
 
             //Se não, Remover do banco
             await _motorcycleRepository.RemoveAsync(motorcycle, cancellationToken);
